pyproject.toml
[project]
name = "get-doxed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "aiogram>=3.19.0",
    "aiohttp>=3.11.16",
    "dishka>=1.5.2",
    "fastapi>=0.115.12",
    "imageio>=2.37.0",
    "jinja2>=3.1.6",
    "motor>=3.7.0",
    "numpy>=2.2.4",
    "pillow>=11.2.0",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "user-agent>=0.1.10",
    "user-agents>=2.2.0",
    "uvicorn>=0.34.0",
]
